chore: lmcut separation procedure SHOULD now be an exact(added epsilon for 0-cost actions), updated lmcut to work with fractional values

+++ code/preprocessing/lmcut.hpp
@@ -12,47 +12,48 @@
 #include "preprocessing.hpp"
 
 namespace lmcut {
-using hmax_function_type = std::function<std::pair<int, int>(const std::vector<unsigned int>&, const std::vector<int>&, const std::vector<int>&)>;
+using hmax_function_type =
+    std::function<std::pair<int, double>(const std::vector<unsigned int>&, const std::vector<double>&, const std::vector<double>&)>;
 
 /**
  * Hmax policy with arbitrary tie-breaking (FCFS)
  */
 [[nodiscard]]
-std::pair<int, int> hmax_arbitrary(const std::vector<unsigned int>& preconditions, const std::vector<int>& hmax_values,
-                                   [[maybe_unused]] const std::vector<int>& _);
+std::pair<int, double> hmax_arbitrary(const std::vector<unsigned int>& preconditions, const std::vector<double>& hmax_values,
+                                      [[maybe_unused]] const std::vector<double>& _);
 
 /**
  * Hmax policy with inverse tie-breaking (LCFS)
  */
 [[nodiscard]]
-std::pair<int, int> hmax_inverse(const std::vector<unsigned int>& preconditions, const std::vector<int>& hmax_values,
-                                 [[maybe_unused]] const std::vector<int>& _);
+std::pair<int, double> hmax_inverse(const std::vector<unsigned int>& preconditions, const std::vector<double>& hmax_values,
+                                    [[maybe_unused]] const std::vector<double>& _);
 
 /**
  * Hmax policy with VDM tie-breaking
  */
 [[nodiscard]]
-std::pair<int, int> hmax_value_decrease_minimization(const std::vector<unsigned int>& preconditions, const std::vector<int>& hmax_values,
-                                                     const std::vector<int>& initial_hmax_values);
+std::pair<int, double> hmax_value_decrease_minimization(const std::vector<unsigned int>& preconditions, const std::vector<double>& hmax_values,
+                                                        const std::vector<double>& initial_hmax_values);
 
 /**
  * Hmax policy with random tie-breaking
  */
 [[nodiscard]]
-std::pair<int, int> hmax_random(const std::vector<unsigned int>& preconditions, const std::vector<int>& hmax_values,
-                                [[maybe_unused]] const std::vector<int>& _);
+std::pair<int, double> hmax_random(const std::vector<unsigned int>& preconditions, const std::vector<double>& hmax_values,
+                                   [[maybe_unused]] const std::vector<double>& _);
 
 /**
  * Initialization function for lmcut
  */
-void init_hmax(const hplus::instance& inst, std::vector<int>& hmax_values, std::vector<int>& pcf, std::vector<int>& pcf_hmax,
-               std::vector<int>& reduced_costs, std::vector<unsigned int>& initial_actions);
+void init_hmax(const hplus::instance& inst, std::vector<double>& hmax_values, std::vector<int>& pcf, std::vector<double>& pcf_hmax,
+               std::vector<double>& reduced_costs, std::vector<unsigned int>& initial_actions);
 
 /**
  * Computes LMcut with the specified tie-breaking policy
  */
-void compute_lmcut(const hplus::instance& inst, std::vector<int> hmax_values, std::vector<int> pcf, std::vector<int> pcf_hmax,
-                   std::vector<int> reduced_costs, const std::vector<unsigned int>& goal_sparse, const std::vector<unsigned int>& initial_actions,
+void compute_lmcut(const hplus::instance& inst, std::vector<double> hmax_values, std::vector<int> pcf, std::vector<double> pcf_hmax,
+                   std::vector<double> reduced_costs, const std::vector<unsigned int>& goal_sparse, const std::vector<unsigned int>& initial_actions,
                    const hmax_function_type& hmax_function, std::vector<std::vector<unsigned int>>& landmarks, bool print = false);
 }  // namespace lmcut
